file check

+++ serial_to_file.py
@@ -4,10 +4,15 @@
 ## requires pySerial to be installed 
 import serial
 import time
+import os
 
 serial_port = '/dev/cu.usbmodem1431';
 baud_rate = 9600; #In arduino, Serial.begin(baud_rate)
 write_to_file_path = "output.txt";
+
+if os.path.isfile(write_to_file_path) :
+    print "output.txt already exists!"
+    exit()
 
 output_file = open(write_to_file_path, "w+");
 ser = serial.Serial(serial_port, baud_rate)
